+++ NEWS.md
@@ -6,6 +6,7 @@
   * File extension can now be set [#170](https://github.com/jmbarbone/mark/issues/170), which by default includes `qmd` ([#163](https://github.com/jmbarbone/mark/issues/163)) and `py` files
   * new parameter `ignore` to ignore any files
   * file paths and line numbers can now be _clicked_ within RStudio ([#171](https://github.com/jmbarbone/mark/issues/171))
+* adds more use of `rlang::list2()` for internally [#199](https://github.com/jmbarbone/mark/issues/199)
 
 # mark 0.6.1